# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment QR Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transaction/{terminal:uuid}/pagination": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Вернуть транзакции пагинацией по терминалу",
                "operationId": "26b6cd545e5cf87d03fdb562aa223a24",
                "parameters": [
                    {
                        "name": "terminal:uuid",
                        "in": "path",
                        "description": "UUID транзакции",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TransactionResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully return transaction pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка возврата пагинации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed return transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transaction/{terminal:uuid}/all": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Вернуть все транзакции у терминал без пагинации",
                "operationId": "5664b2347060f839cfc52c7710159ee4",
                "parameters": [
                    {
                        "name": "terminal:uuid",
                        "in": "path",
                        "description": "UUID транзакции",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TransactionResource"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get all transaction by terminal"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transaction/{terminal:uuid}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Вернуть транзакцию по uuid",
                "operationId": "2ebe09b4d2a6a42b8cf1efdbf28e5fcc",
                "parameters": [
                    {
                        "name": "terminal:uuid",
                        "in": "path",
                        "description": "UUID транзакции",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TransactionResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Show transaction by uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transaction": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Вернуть транзакцию по uuid",
                "operationId": "f48ce787ab516614850907c258b918aa",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "amount": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "terminal_uuid": {
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "По uuid такой записи terminala не сущесвует.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Такой записи по uuid не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TransactionResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully create transaction"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TransactionResource": {
                "title": "Transaction Resource",
                "properties": {
                    "driver_currency_id": {
                        "type": "integer"
                    },
                    "amount": {
                        "type": "string"
                    },
                    "uuid": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "code": {
                        "type": "integer",
                        "example": "500"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}